# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SEMESTA ISLAM Developer API
  description: |
    Canonical Developer Platform API specification for SEMESTA ISLAM — a trusted platform connecting learners with verified Islamic educators.
    Governed by docs/contracts/ and docs/07_API_ENDPOINTS.md.
    Contract semantics per Founder Decisions #1–#3 (2026-08-01): sanad fields domain-neutral; verifier authorization is payload-canonical (`verifierRoles`); 2xx/error envelope is runtime-canonical (see docs/audit/CONTRACT_DRIFT_REPORT.md §4).
  version: 1.0.0
  contact:
    name: SEMESTA ISLAM Developer Operations
    url: https://semestaislam.com/developers
servers:
  - url: http://localhost:3000
    description: Localhost Development Server (Local Demo Mode)
  - url: https://sandbox.semestaislam.com
    description: SEMESTA Sandbox (Mock Infrastructure)

paths:
  /api/v1/bookings/inquire:
    post:
      summary: Submit Booking Inquiry [IMPLEMENTED + RUNTIME VERIFIED]
      description: Submits a booking inquiry for an educator and logs an initial points accrual event.
      operationId: submitBookingInquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - educatorId
                - learningMethod
                - preferredSchedule
                - learnerName
                - contactPhone
              properties:
                educatorId:
                  type: string
                  format: uuid
                courseId:
                  type: string
                  format: uuid
                scheduleId:
                  type: string
                  format: uuid
                learningMethod:
                  type: string
                  enum: [ONLINE_ZOOM, PRIVATE_HOME, GROUP_MAJELIS]
                preferredSchedule:
                  type: string
                  minLength: 3
                learnerName:
                  type: string
                  minLength: 2
                contactPhone:
                  type: string
                  pattern: "^(\\+62|08)[0-9]{8,12}$"
                notes:
                  type: string
                  maxLength: 500
      responses:
        '201':
          description: Booking inquiry successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation Failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /api/v1/bookings/confirm:
    post:
      summary: Confirm Pending Booking [IMPLEMENTED + RUNTIME VERIFIED]
      description: Educator/Founder confirms a PENDING booking. Authorization is MVP payload-canonical via `actorRoles`; PHASE 2 will replace with session-based identity + RBAC.
      operationId: confirmBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - bookingId
                - actorUserId
                - actorRoles
              properties:
                bookingId:
                  type: string
                  format: uuid
                actorUserId:
                  type: string
                  format: uuid
                actorRoles:
                  type: array
                  items:
                    type: string
                    enum: [LEARNER, GUARDIAN, EDUCATOR, INSTITUTION_ADMIN, LAJNAH_VERIFIER, FOUNDER_ADMIN]
      responses:
        '200':
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation Failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Conflict - Booking not in PENDING state (e.g. re-confirm of CONFIRMED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /api/v1/verification/submit:
    post:
      summary: Submit Verification Request [IMPLEMENTED + RUNTIME VERIFIED]
      description: Submits educator credential documents and SHA-256 fingerprint for Lajnah review.
      operationId: submitVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - educatorId
                - ktpNumber
                - ktpDocumentUrl
                - ijazahDocumentUrl
                - ijazahSha256Hash
                - recommenderEmail
                - recommenderInstitution
              properties:
                educatorId:
                  type: string
                  format: uuid
                ktpNumber:
                  type: string
                  minLength: 16
                  maxLength: 16
                ktpDocumentUrl:
                  type: string
                  format: uri
                ijazahDocumentUrl:
                  type: string
                  format: uri
                ijazahSha256Hash:
                  type: string
                  pattern: "^[a-fA-F0-9]{64}$"
                recommenderEmail:
                  type: string
                  format: email
                recommenderInstitution:
                  type: string
                  minLength: 3
                qiraahSanadName:
                  type: string
      responses:
        '201':
          description: Verification request submitted to queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Invalid SHA-256 hash or payload validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /api/v1/verification/review:
    post:
      summary: Evaluate Verification Review [IMPLEMENTED + RUNTIME VERIFIED]
      description: |
        Evaluates state machine transition for Lajnah verifiers.
        MVP authorization is PAYLOAD-CANONICAL: the `verifierRoles` array in the JSON body gates access (LAJNAH_VERIFIER / FOUNDER_ADMIN → allowed; others → 403).
        No header-based auth (`X-Verifier-Role`) is enforced at runtime; `X-Verifier-Role` remains a `[FUTURE PROPOSAL]`. PHASE 2 replaces payload roles with session identity + RBAC.
      operationId: reviewVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - verificationRequestId
                - verifierUserId
                - verifierRoles
                - currentStatus
                - targetStatus
                - reviewNotes
              properties:
                verificationRequestId:
                  type: string
                verifierUserId:
                  type: string
                  format: uuid
                verifierRoles:
                  type: array
                  items:
                    type: string
                    enum: [LEARNER, GUARDIAN, EDUCATOR, INSTITUTION_ADMIN, LAJNAH_VERIFIER, FOUNDER_ADMIN]
                currentStatus:
                  type: string
                  enum: [DRAFT, SUBMITTED, UNDER_REVIEW_LAJNAH, VERIFIED, REJECTED, REVOKED]
                targetStatus:
                  type: string
                  enum: [DRAFT, SUBMITTED, UNDER_REVIEW_LAJNAH, VERIFIED, REJECTED, REVOKED]
                reviewNotes:
                  type: string
                  minLength: 5
                ethicsScore:
                  type: number
                  minimum: 0
                  maximum: 100
      responses:
        '200':
          description: Verification status successfully transitioned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '403':
          description: Forbidden - Insufficient verifier role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Conflict - Illegal state machine transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /api/v1/verification/resubmit:
    post:
      summary: Resubmit Rejected Verification [IMPLEMENTED + RUNTIME VERIFIED]
      description: Resubmits a rejected verification back to SUBMITTED status.
      operationId: resubmitVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - verificationRequestId
                - educatorId
                - ktpNumber
                - ktpDocumentUrl
                - ijazahDocumentUrl
                - ijazahSha256Hash
                - recommenderEmail
                - recommenderInstitution
              properties:
                verificationRequestId:
                  type: string
                currentStatus:
                  type: string
                  enum: [REJECTED, REVOKED]
      responses:
        '200':
          description: Resubmission accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '409':
          description: Invalid transition state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /api/v1/verification/status:
    get:
      summary: Get Verification Status [IMPLEMENTED + RUNTIME VERIFIED]
      description: Retrieves the current verification status and review timeline for an educator.
      operationId: getVerificationStatus
      parameters:
        - name: educatorId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Verification status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation failure (invalid educatorId)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: No verification request found for this educator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /api/v1/educators:
    get:
      summary: List Public Educators [CONTRACT ONLY — FUTURE / ASPIRATIONAL]
      description: |
        Query verified educators with filtering by teaching method and location.
        [CONTRACT ONLY] No runtime GET list API exists in the B2C MVP. The `meta`/`PaginationMeta` envelope is `[FUTURE / ASPIRATIONAL]` and not required for current MVP endpoints.
      operationId: listEducators
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: method
          in: query
          schema:
            type: string
            enum: [all, ONLINE_ZOOM, PRIVATE_HOME, GROUP_MAJELIS]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of public educator profiles
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Educator'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

  /api/v1/educators/{id}:
    get:
      summary: Get Educator Detail [CONTRACT ONLY]
      description: Retrieve public profile, sanad records, and trust metadata for a specific educator.
      operationId: getEducatorById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Educator detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Educator'
        '404':
          description: Educator not found

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: >-
        [FUTURE PROPOSAL — DEFERRED] Bearer API Key authentication for developers.
        Not enforced by any runtime route in the B2C MVP. Belongs to the post-MVP Developer Platform.
    VerifierRoleAuth:
      type: apiKey
      in: header
      name: X-Verifier-Role
      description: >-
        [FUTURE PROPOSAL — NOT IMPLEMENTED] Lajnah Verifier authorization header.
        The B2C MVP authorizes via the `verifierRoles` array in the JSON payload (Decision #2 — payload-canonical).
        No runtime route reads this header.

  schemas:
    SuccessEnvelope:
      type: object
      description: Canonical 2xx response envelope (Decision #3 — runtime-canonical)
      properties:
        success:
          type: boolean
          example: true
        statusCode:
          type: integer
          example: 200
        message:
          type: string
        data:
          description: Endpoint-specific payload
      required: [success, statusCode, message]
    Educator:
      $ref: './schemas/educator.schema.json'
    Course:
      $ref: './schemas/course.schema.json'
    Sanad:
      $ref: './schemas/sanad.schema.json'
    Verification:
      $ref: './schemas/verification.schema.json'
    TrustMetadata:
      $ref: './schemas/trust-metadata.schema.json'
    ErrorEnvelope:
      $ref: './schemas/error.schema.json'
    PaginationMeta:
      $ref: './schemas/pagination.schema.json'
    WebhookEvent:
      $ref: './schemas/webhook-event.schema.json'